Split views.py into seperate files
