+++ infraohjelmointi_api/views/BaseViewSet.py
@@ -1,0 +1,11 @@
+from rest_framework import viewsets
+from overrides import override
+
+
+class BaseViewSet(viewsets.ModelViewSet):
+    @override
+    def get_queryset(self):
+        """
+        Overriden ModelViewSet class method to get appropriate queryset using serializer class
+        """
+        return self.get_serializer_class().Meta.model.objects.all()
