+++ infraohjelmointi_api/views/ConstructionPhaseDetailViewSet.py
@@ -1,0 +1,13 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.ConstructionPhaseDetailSerializer import (
+    ConstructionPhaseDetailSerializer,
+)
+
+
+class ConstructionPhaseDetailViewSet(BaseViewSet):
+    """
+    API endpoint that allows construction phase details to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ConstructionPhaseDetailSerializer
