+++ infraohjelmointi_api/views/ConstructionPhaseViewSet.py
@@ -1,0 +1,13 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.ConstructionPhaseSerializer import (
+    ConstructionPhaseSerializer,
+)
+
+
+class ConstructionPhaseViewSet(BaseViewSet):
+    """
+    API endpoint that allows Construction phases to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ConstructionPhaseSerializer
