+++ infraohjelmointi_api/views/MockProjectViewSet.py
@@ -1,0 +1,17 @@
+import json
+from rest_framework import viewsets
+from rest_framework.response import Response
+
+
+class MockProjectViewSet(viewsets.ViewSet):
+    """
+    API endpoint that returns mock project data.
+    """
+
+    mock_data = json.load(
+        open("./infraohjelmointi_api/mock_data/hankekortti.json", "r")
+    )
+
+    def list(self, request):
+        queryset = self.mock_data
+        return Response(queryset)
