+++ infraohjelmointi_api/views/PlanningPhaseViewSet.py
@@ -1,0 +1,13 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.PlanningPhaseSerializer import (
+    PlanningPhaseSerializer,
+)
+
+
+class PlanningPhaseViewSet(BaseViewSet):
+    """
+    API endpoint that allows Planning phases to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = PlanningPhaseSerializer
