+++ infraohjelmointi_api/views/ProjectCategoryViewSet.py
@@ -1,0 +1,13 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.ProjectCategorySerializer import (
+    ProjectCategorySerializer,
+)
+
+
+class ProjectCategoryViewSet(BaseViewSet):
+    """
+    API endpoint that allows project cetagories to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ProjectCategorySerializer
