+++ infraohjelmointi_api/views/ProjectLockViewSet.py
@@ -1,0 +1,11 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.ProjectLockSerializer import ProjectLockSerializer
+
+
+class ProjectLockViewSet(BaseViewSet):
+    """
+    API endpoint that allows Project Lock status to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ProjectLockSerializer
