+++ infraohjelmointi_api/views/ProjectPhaseViewSet.py
@@ -1,0 +1,13 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.ProjectPhaseSerializer import (
+    ProjectPhaseSerializer,
+)
+
+
+class ProjectPhaseViewSet(BaseViewSet):
+    """
+    API endpoint that allows project phase to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ProjectPhaseSerializer
