+++ infraohjelmointi_api/views/ProjectPriorityViewSet.py
@@ -1,0 +1,13 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.ProjectPrioritySerializer import (
+    ProjectPrioritySerializer,
+)
+
+
+class ProjectPriorityViewSet(BaseViewSet):
+    """
+    API endpoint that allows project Priority to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ProjectPrioritySerializer
