+++ infraohjelmointi_api/views/ProjectQualityLevelViewSet.py
@@ -1,0 +1,13 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.ProjectQualityLevelSerializer import (
+    ProjectQualityLevelSerializer,
+)
+
+
+class ProjectQualityLevelViewSet(BaseViewSet):
+    """
+    API endpoint that allows Project quality levels to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ProjectQualityLevelSerializer
