+++ infraohjelmointi_api/views/ProjectResponsibleZoneViewSet.py
@@ -1,0 +1,14 @@
+from .BaseViewSet import BaseViewSet
+
+from infraohjelmointi_api.serializers.ProjectResponsibleZoneSerializer import (
+    ProjectResponsibleZoneSerializer,
+)
+
+
+class ProjectResponsibleZoneViewSet(BaseViewSet):
+    """
+    API endpoint that allows Planning responsible zones to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = ProjectResponsibleZoneSerializer
