+++ infraohjelmointi_api/views/ProjectSetViewSet.py
@@ -1,0 +1,26 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers import (
+    ProjectSetGetSerializer,
+    ProjectSetCreateSerializer,
+)
+
+from overrides import override
+
+
+class ProjectSetViewSet(BaseViewSet):
+    """
+    API endpoint that allows project sets to be viewed or edited.
+    """
+
+    permission_classes = []
+
+    @override
+    def get_serializer_class(self):
+        """
+        Overriden ModelViewSet class method to get appropriate serializer depending on the request action
+        """
+        if self.action == "list":
+            return ProjectSetGetSerializer
+        if self.action == "retrieve":
+            return ProjectSetGetSerializer
+        return ProjectSetCreateSerializer
