+++ infraohjelmointi_api/views/ProjectViewSet.py
@@ -1,569 +1,41 @@
-import time
-import uuid
 from datetime import date
+from django_filters.rest_framework import DjangoFilterBackend
+import django_filters
+from infraohjelmointi_api.serializers import (
+    ProjectHashtagSerializer,
+    ProjectPhaseSerializer,
+    ProjectGetSerializer,
+    ProjectCreateSerializer,
+    ProjectFinancialSerializer,
+    ProjectGroupSerializer,
+    SearchResultSerializer,
+    ProjectNoteGetSerializer,
+)
 from infraohjelmointi_api.models import (
     Project,
-    ProjectClass,
     ProjectGroup,
+    ProjectClass,
     ProjectLocation,
-    ProjectFinancial,
-    ClassFinancial,
 )
-from infraohjelmointi_api.services.ProjectFinancialService import (
+from infraohjelmointi_api.services import (
+    ProjectWiseService,
     ProjectFinancialService,
+    ProjectClassService,
 )
-from infraohjelmointi_api.services.ClassFinancialService import (
-    ClassFinancialService,
-)
-from rest_framework.exceptions import APIException, ParseError, ValidationError
-from django.dispatch import receiver
-import django_filters
-from django.db.models import Q
-from django.db.models.expressions import RawSQL
-from distutils.util import strtobool
-import datetime
-from rest_framework.pagination import PageNumberPagination
-from itertools import chain
-from django.http.response import StreamingHttpResponse
-from django.db.models.signals import post_save
-from rest_framework import viewsets
-from .serializers import (
-    NoteCreateSerializer,
-    NoteHistorySerializer,
-    ProjectCreateSerializer,
-    ProjectGetSerializer,
-    ProjectNoteGetSerializer,
-    ProjectSetGetSerializer,
-    ProjectSetCreateSerializer,
-    ProjectTypeSerializer,
-    PersonSerializer,
-    ProjectAreaSerializer,
-    BudgetItemSerializer,
-    TaskSerializer,
-    ProjectPhaseSerializer,
-    ProjectPrioritySerializer,
-    TaskStatusSerializer,
-    ConstructionPhaseDetailSerializer,
-    ProjectCategorySerializer,
-    ProjectRiskSerializer,
-    NoteGetSerializer,
-    ConstructionPhaseSerializer,
-    PlanningPhaseSerializer,
-    NoteUpdateSerializer,
-    ProjectQualityLevelSerializer,
-    ProjectLocationSerializer,
-    ProjectClassSerializer,
-    ProjectResponsibleZoneSerializer,
-    ProjectHashtagSerializer,
-    ProjectGroupSerializer,
-    ProjectLockSerializer,
-    SearchResultSerializer,
-    ProjectFinancialSerializer,
-    ClassFinancialSerializer,
-)
-from .paginations import StandardResultsSetPagination
-from .services import ProjectClassService, ProjectLocationService, ProjectWiseService
-from rest_framework import status
-from rest_framework.views import APIView
-from rest_framework.response import Response
 import json
-from rest_framework import status
-from rest_framework.decorators import action
-
-from django.db import transaction
+from .BaseViewSet import BaseViewSet
+from distutils.util import strtobool
+from ..paginations import StandardResultsSetPagination
 from overrides import override
-from django_filters.rest_framework import DjangoFilterBackend
-from django.db.models import Count, Case, When
-from django.shortcuts import get_object_or_404
+from rest_framework.response import Response
 from django.db import transaction
-import logging
-
-
-logger = logging.getLogger("infraohjelmointi_api")
-
-
-class BaseViewSet(viewsets.ModelViewSet):
-    @override
-    def get_queryset(self):
-        """
-        Overriden ModelViewSet class method to get appropriate queryset using serializer class
-        """
-        return self.get_serializer_class().Meta.model.objects.all()
-
-
-class ProjectFinancialViewSet(BaseViewSet):
-    """
-    API endpoint that allows Project Finances to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectFinancialSerializer
-
-    @action(
-        methods=["get"],
-        detail=False,
-        url_path=r"(?P<project>[0-9a-f]{8}\-[0-9a-f]{4}\-4[0-9a-f]{3}\-[89ab][0-9a-f]{3}\-[0-9a-f]{12})/(?P<year>[0-9]{4})",
-    )
-    def get_finances_by_year(self, request, project, year):
-        """
-        Custom action to get finances of a project for an year.
-
-            URL Parameters
-            ----------
-
-            project_id : UUID string
-            year : Int
-
-            Usage
-            ----------
-
-            project-financials/<project_id>/<year>
-
-            Returns
-            -------
-
-            JSON
-                ProjectFinancial instance for the given year and project id
-        """
-        queryFilter = {"project": project, "year": year}
-        finance_object = get_object_or_404(ProjectFinancial, **queryFilter)
-        return Response(ProjectFinancialSerializer(finance_object).data)
-
-
-class ProjectLockViewSet(BaseViewSet):
-    """
-    API endpoint that allows Project Lock status to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectLockSerializer
-
-
-class ProjectHashtagViewSet(BaseViewSet):
-    """
-    API endpoint that allows Project Hashtags to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectHashtagSerializer
-
-    @override
-    def list(self, request, *args, **kwargs):
-        """
-        Overriden list action to get custom response for a GET request to hashtags endpoint
-
-            Usage
-            ----------
-
-            project-hashtags/
-
-            Returns
-            -------
-
-            List of ProjectHashtag and top 15 most used hash tags
-
-            JSON
-                { "hashTags": [ProjectHashTag], "popularHashTags": [ProjectHashTag] }
-        """
-        qs = self.get_queryset().prefetch_related("relatedProject")
-        popularQs = (
-            qs.annotate(usage_count=Count("relatedProject"))
-            .filter(usage_count__gt=0)
-            .order_by("-usage_count")[:15]
-        )
-
-        serializer = self.get_serializer(qs, many=True)
-        popularSerializer = self.get_serializer(popularQs, many=True)
-        return Response(
-            {"hashTags": serializer.data, "popularHashTags": popularSerializer.data}
-        )
-
-
-class ProjectGroupViewSet(BaseViewSet):
-    """
-    API endpoint that allows Project Groups to be viewed or edited.
-    """
-
-    serializer_class = ProjectGroupSerializer
-
-    @override
-    def list(self, request, *args, **kwargs):
-        """
-        Overriden list action to get a list of ProjectGroup
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Groups with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-groups/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectGroup instances with financial sums for projects under each group
-        """
-        year = request.query_params.get("year", date.today().year)
-        qs = self.get_queryset()
-        serializer = self.get_serializer(qs, many=True, context={"finance_year": year})
-
-        return Response(serializer.data)
-
-    @override
-    def destroy(self, request, *args, **kwargs):
-        """
-        Overriding destroy action to get the deleted group id as a response
-        """
-        group = self.get_object()
-        data = group.id
-        group.delete()
-        return Response({"id": data})
-
-    @action(
-        methods=["get"],
-        detail=False,
-        url_path=r"coordinator",
-    )
-    def get_groups_for_coordinator(self, request):
-        """
-        Custom action to get ProjectGroup instances with coordinator location/classes
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Groups with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-groups/coordinator/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectGroup instances with financial sums for projects under each group
-        """
-        year = request.query_params.get("year", date.today().year)
-        qs = self.get_queryset().select_related(
-            "classRelation",
-            "locationRelation",
-            "classRelation__coordinatorClass",
-            "locationRelation__coordinatorLocation",
-            "classRelation__parent__coordinatorClass",
-            "locationRelation__parent__coordinatorLocation",
-            "locationRelation__parent__parent__coordinatorLocation",
-        )
-        serializer = self.get_serializer(
-            qs, many=True, context={"finance_year": year, "for_coordinator": True}
-        )
-        return Response(serializer.data)
-
-    permission_classes = []
-    serializer_class = ProjectGroupSerializer
-
-
-class ProjectLocationViewSet(BaseViewSet):
-    """
-    API endpoint that allows Project Locations to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectLocationSerializer
-
-    @override
-    def list(self, request, *args, **kwargs):
-        """
-        Overriden list action to get a list of ProjectLocation
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Locations with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-locations/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectLocation instances with financial sums for projects under each location
-        """
-        year = request.query_params.get("year", date.today().year)
-        qs = self.get_queryset()
-        serializer = self.get_serializer(qs, many=True, context={"finance_year": year})
-
-        return Response(serializer.data)
-
-    def get_queryset(self):
-        """Default is programmer view"""
-        return ProjectLocationService.list_all()
-
-    @action(methods=["get"], detail=False, url_path=r"coordinator")
-    def list_for_coordinator(self, request):
-        """
-        Overriden list action to get a list of coordinator ProjectLocations
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Locations with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-locations/coordinator/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectLocation instances with financial sums for projects under each location
-        """
-        year = request.query_params.get("year", date.today().year)
-        serializer = ProjectLocationSerializer(
-            ProjectLocationService.list_all_for_coordinator(),
-            many=True,
-            context={
-                "finance_year": year,
-                "for_coordinator": True,
-            },
-        )
-        return Response(serializer.data)
-
-
-class ProjectClassViewSet(BaseViewSet):
-    """
-    API endpoint that allows Project Classes to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectClassSerializer
-
-    @override
-    def list(self, request, *args, **kwargs):
-        """
-        Overriden list action to get a list of ProjectClass
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Class with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-classes/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectClass instances with financial sums for projects under each class
-        """
-        year = request.query_params.get("year", date.today().year)
-        qs = self.get_queryset()
-        serializer = self.get_serializer(qs, many=True, context={"finance_year": year})
-
-        return Response(serializer.data)
-
-    @override
-    def get_queryset(self):
-        """Default is programmer view"""
-        return (
-            ProjectClassService.list_all()
-            .select_related("coordinatorClass")
-            .prefetch_related("coordinatorClass__finances")
-        )
-
-    @action(methods=["get"], detail=False, url_path=r"coordinator")
-    def list_for_coordinator(self, request):
-        """
-        Overriden list action to get a list of coordinator ProjectClass
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Class with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-classes/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectClass instances with financial sums for projects under each class
-        """
-        year = request.query_params.get("year", date.today().year)
-        serializer = ProjectClassSerializer(
-            ProjectClassService.list_all_for_coordinator()
-            .prefetch_related("coordinatorClass__finances")
-            .select_related("coordinatorClass"),
-            many=True,
-            context={
-                "finance_year": year,
-                "for_coordinator": True,
-            },
-        )
-        return Response(serializer.data)
-
-    def is_patch_data_valid(self, data):
-        """
-        Utility function to validate patch data sent to the custom PATCH endpoint for coordinator class finances
-        """
-        finances = data.get("finances", None)
-        if finances == None:
-            return False
-
-        parameters = finances.keys()
-        if "year" not in parameters:
-            return False
-
-        for param in parameters:
-            if param == "year":
-                continue
-            values = finances[param]
-            values_length = len(values.keys())
-
-            if type(values) != dict:
-                return False
-            if values_length == 0 or values_length > 2:
-                return False
-            if not "frameBudget" in values and not "budgetChange" in values:
-                return False
-
-        return True
-
-    @action(
-        methods=["patch"],
-        detail=False,
-        url_path=r"coordinator/(?P<class_id>[0-9a-f]{8}\-[0-9a-f]{4}\-4[0-9a-f]{3}\-[89ab][0-9a-f]{3}\-[0-9a-f]{12})",
-    )
-    def patch_coordinator_class_finances(self, request, class_id):
-        """
-        Custom PATCH endpoint for coordinator classes **finances ONLY**
-
-            URL Parameters
-            ----------
-
-            class_id : UUID string
-
-            Coordinator class Id
-
-            Usage
-            ----------
-
-            project-classes/coordinator/<class_id>/
-
-            Returns
-            -------
-
-            JSON
-                Patched coordinator ProjectClass Instance
-        """
-        if not ProjectClassService.instance_exists(
-            id=class_id, forCoordinatorOnly=True
-        ):
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
-        if not self.is_patch_data_valid(request.data):
-            return Response(
-                data={"message": "Invalid data format"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        finances = request.data.get("finances")
-        startYear = finances.get("year")
-        for parameter in finances.keys():
-            if parameter == "year":
-                continue
-            patchData = finances[parameter]
-            year = ClassFinancialService.get_request_field_to_year_mapping(
-                start_year=startYear
-            ).get(parameter, None)
-
-            if year == None:
-                return Response(
-                    data={"message": "Invalid data format"},
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-
-        ClassFinancialService.update_or_create(
-            year=year, class_id=class_id, updatedData=patchData
-        )
-        return Response(
-            ProjectClassSerializer(
-                ProjectClassService.get_by_id(id=class_id),
-                context={
-                    "finance_year": startYear,
-                    "for_coordinator": True,
-                },
-            ).data
-        )
-
-
-class ProjectQualityLevelViewSet(BaseViewSet):
-    """
-    API endpoint that allows Project quality levels to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectQualityLevelSerializer
-
-
-class PlanningPhaseViewSet(BaseViewSet):
-    """
-    API endpoint that allows Planning phases to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = PlanningPhaseSerializer
-
-
-class ProjectResponsibleZoneViewSet(BaseViewSet):
-    """
-    API endpoint that allows Planning responsible zones to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectResponsibleZoneSerializer
-
-
-class ConstructionPhaseViewSet(BaseViewSet):
-    """
-    API endpoint that allows Construction phases to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ConstructionPhaseSerializer
+from rest_framework.decorators import action
+from rest_framework.exceptions import APIException, ParseError, ValidationError
+from rest_framework.pagination import PageNumberPagination
+import uuid
+from rest_framework import status
+from itertools import chain
+from django.db.models import Count, Case, When, Q
 
 
 class ProjectFilter(django_filters.FilterSet):
@@ -1651,242 +1123,3 @@ class ProjectViewSet(BaseViewSet):
             qs = qs.filter(Q(id__in=financialProjectIds) & Q(programmed=True))
 
         return qs
-
-
-class TaskStatusViewSet(BaseViewSet):
-    """
-    API endpoint that allows project types to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = TaskStatusSerializer
-
-
-class ProjectTypeViewSet(BaseViewSet):
-    """
-    API endpoint that allows project types to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectTypeSerializer
-
-
-class ConstructionPhaseDetailViewSet(BaseViewSet):
-    """
-    API endpoint that allows construction phase details to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ConstructionPhaseDetailSerializer
-
-
-class ProjectCategoryViewSet(BaseViewSet):
-    """
-    API endpoint that allows project cetagories to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectCategorySerializer
-
-
-class ProjectRiskViewSet(BaseViewSet):
-    """
-    API endpoint that allows project risk assessments to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectRiskSerializer
-
-
-class ProjectPhaseViewSet(BaseViewSet):
-    """
-    API endpoint that allows project phase to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectPhaseSerializer
-
-
-class ProjectPriorityViewSet(BaseViewSet):
-    """
-    API endpoint that allows project Priority to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectPrioritySerializer
-
-
-class MockProjectViewSet(viewsets.ViewSet):
-    """
-    API endpoint that returns mock project data.
-    """
-
-    mock_data = json.load(
-        open("./infraohjelmointi_api/mock_data/hankekortti.json", "r")
-    )
-
-    def list(self, request):
-        queryset = self.mock_data
-        return Response(queryset)
-
-
-class PersonViewSet(BaseViewSet):
-    """
-    API endpoint that allows persons to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = PersonSerializer
-
-
-class ProjectSetViewSet(BaseViewSet):
-    """
-    API endpoint that allows project sets to be viewed or edited.
-    """
-
-    permission_classes = []
-
-    @override
-    def get_serializer_class(self):
-        """
-        Overriden ModelViewSet class method to get appropriate serializer depending on the request action
-        """
-        if self.action == "list":
-            return ProjectSetGetSerializer
-        if self.action == "retrieve":
-            return ProjectSetGetSerializer
-        return ProjectSetCreateSerializer
-
-
-class ProjectAreaViewSet(BaseViewSet):
-    """
-    API endpoint that allows project areas to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ProjectAreaSerializer
-
-
-class BudgetItemViewSet(BaseViewSet):
-    """
-    API endpoint that allows Budgets to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = BudgetItemSerializer
-
-
-class TaskViewSet(BaseViewSet):
-    """
-    API endpoint that allows Tasks to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = TaskSerializer
-
-
-class NoteViewSet(BaseViewSet):
-
-    """
-    API endpoint that allows notes to be viewed or edited.
-    """
-
-    permission_classes = []
-
-    @override
-    def get_serializer_class(self):
-        """
-        Overriden ModelViewSet class method to get appropriate serializer depending on the request action
-        """
-        if self.action in ["list", "retrieve"]:
-            return NoteGetSerializer
-        if self.action == "create":
-            return NoteCreateSerializer
-        return NoteUpdateSerializer
-
-    @action(methods=["get"], detail=True, url_path=r"history")
-    def history(self, request, pk):
-        """
-        Custom action to get edit history of a note
-
-            URL Parameters
-            ----------
-
-            note_id : UUID string
-
-            Usage
-            ----------
-
-            notes/<note_id>/history/
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectNote instances
-        """
-        try:
-            uuid.UUID(str(pk))  # validating UUID
-            instance = self.get_object()
-            qs = instance.history.all()
-            serializer = NoteHistorySerializer(qs, many=True)
-            return Response(serializer.data)
-        except ValueError:
-            return Response(
-                data={"message": "Invalid UUID"}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-    @override
-    def destroy(self, request, *args, **kwargs):
-        """
-        Overriding destroy action to soft delete note on DELETE request
-        """
-        note = self.get_object()
-        data = note.id
-        note.deleted = True
-        note.save()
-        return Response({"id": data})
-
-    @action(methods=["get"], detail=True, url_path=r"history/(?P<userId>[-\w]+)")
-    def history_user(self, request, pk, userId):
-        """
-        Custom action to get edit history of a note edited by a specific user
-            URL Parameters
-            ----------
-
-            note_id : UUID string
-
-            user_id : UUID string
-
-            Usage
-            ----------
-
-            notes/<note_id>/history/<user_id>/
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectNote instances
-        """
-        try:
-            uuid.UUID(str(userId))
-            uuid.UUID(str(pk))
-            instance = self.get_object()
-            qs = instance.history.all().filter(updatedBy_id=userId)
-            serializer = NoteHistorySerializer(qs, many=True)
-            return Response(serializer.data)
-
-        except ValueError:
-            return Response(
-                data={"message": "Invalid UUID"}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-
-class ClassFinancialViewSet(BaseViewSet):
-    """
-    API endpoint that allows Class Financials to be viewed or edited.
-    """
-
-    permission_classes = []
-    serializer_class = ClassFinancialSerializer
