+++ infraohjelmointi_api/views/TaskStatusViewSet.py
@@ -1,0 +1,11 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.TaskStatusSerializer import TaskStatusSerializer
+
+
+class TaskStatusViewSet(BaseViewSet):
+    """
+    API endpoint that allows project types to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = TaskStatusSerializer
