+++ infraohjelmointi_api/views/TaskViewSet.py
@@ -1,0 +1,11 @@
+from .BaseViewSet import BaseViewSet
+from infraohjelmointi_api.serializers.TaskSerializer import TaskSerializer
+
+
+class TaskViewSet(BaseViewSet):
+    """
+    API endpoint that allows Tasks to be viewed or edited.
+    """
+
+    permission_classes = []
+    serializer_class = TaskSerializer
