+++ infraohjelmointi_api/views/__init__.py
@@ -1,0 +1,27 @@
+from .BaseViewSet import *
+from .ProjectFinancialViewSet import *
+from .ProjectLockViewSet import *
+from .ProjectHashtagViewSet import *
+from .ProjectGroupViewSet import *
+from .ProjectLocationViewSet import *
+from .ProjectClassViewSet import *
+from .ProjectQualityLevelViewSet import *
+from .PlanningPhaseViewSet import *
+from .ProjectResponsibleZoneViewSet import *
+from .ConstructionPhaseViewSet import *
+from .ProjectViewSet import *
+from .TaskStatusViewSet import *
+from .ProjectTypeViewSet import *
+from .ConstructionPhaseDetailViewSet import *
+from .ProjectCategoryViewSet import *
+from .ProjectRiskViewSet import *
+from .ProjectPhaseViewSet import *
+from .ProjectPriorityViewSet import *
+from .MockProjectViewSet import *
+from .PersonViewSet import *
+from .ProjectSetViewSet import *
+from .ProjectAreaViewSet import *
+from .BudgetItemViewSet import *
+from .TaskViewSet import *
+from .NoteViewSet import *
+from .ClassFinancialViewSet import *
